Fix Vercel deployment failure: isolate getDataSyncDate from species-store.ts
The /api/data-sync-date function was 427.68MB uncompressed, exceeding
Vercel's 250MB limit. species-store.ts reads data/redlist etc. via
computed paths that Next.js's file-tracer can't statically resolve, so
it conservatively bundled the entire ~400MB data/ directory into any
function importing anything from that module — including this one,
which only ever needed to read latest-sync.txt.

Moved getDataSyncDate() into its own zero-dependency module
(src/lib/data-sync-date.ts) with no import from species-store.ts.
Confirmed via the route's .next trace manifest: 427MB → 1.16MB, 0
files under data/ now traced for this route (down from the full
dataset), while other routes that genuinely need the data (e.g.
/api/redlist/taxa-summary) are unaffected.
