+++ app/src/lib/data/species-store.ts
@@ -18,20 +18,6 @@ const DATA_DIR = path.join(process.cwd(), "data");
 const REDLIST_DIR = path.join(DATA_DIR, "redlist");
 const TAXA_SUMMARY_PATH = path.join(DATA_DIR, "taxa-summary.json");
 const NODE_CHILDREN_SUMMARIES_PATH = path.join(DATA_DIR, "node-children-summaries.json");
-const LATEST_SYNC_PATH = path.join(process.cwd(), "latest-sync.txt");
-
-/**
- * The timestamp of the data sync that data/taxa-summary.json etc. were built
- * from (see README § Data Sync Pipeline) — i.e. the "now" the build script used
- * when it computed isOutdated() into that static file. Parses
- * scripts/upload-data-to-r2.ts's makeTimestamp format ("2026-07-16T17-03-30Z",
- * colons swapped for dashes to be S3/URL-safe) back into a Date.
- */
-export function getDataSyncDate(): Date {
-  const raw = fs.readFileSync(LATEST_SYNC_PATH, "utf-8").trim();
-  const iso = raw.replace(/T(\d{2})-(\d{2})-(\d{2})Z$/, "T$1:$2:$3Z");
-  return new Date(iso);
-}
 
 // =============================================================================
 // TYPES
